+++ cmd/s3-gw/app_metrics.go
@@ -1,8 +1,13 @@
 package main
 
 import (
+	"context"
 	"net/http"
+	"strings"
+	"time"
 
+	"github.com/nspcc-dev/neofs-sdk-go/client"
+	"github.com/nspcc-dev/neofs-sdk-go/pool"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/spf13/viper"
@@ -12,6 +17,8 @@ import (
 const (
 	namespace      = "neofs_s3_gw"
 	stateSubsystem = "state"
+
+	checkNeoFSConnectionTimeOut = 5 * time.Second
 )
 
 type GateMetrics struct {
@@ -19,13 +26,14 @@ type GateMetrics struct {
 }
 
 type stateMetrics struct {
+	p           *pool.Pool
 	healthCheck prometheus.Gauge
 	gwVersion   *prometheus.GaugeVec
 }
 
-func newGateMetrics() *GateMetrics {
-	stateMetric := newStateMetrics()
-	stateMetric.register()
+func newGateMetrics(p *pool.Pool) *GateMetrics {
+	stateMetric := newStateMetrics(p)
+	prometheus.MustRegister(stateMetric)
 
 	return &GateMetrics{
 		stateMetrics: *stateMetric,
@@ -36,8 +44,9 @@ func (g *GateMetrics) Unregister() {
 	g.stateMetrics.unregister()
 }
 
-func newStateMetrics() *stateMetrics {
+func newStateMetrics(p *pool.Pool) *stateMetrics {
 	return &stateMetrics{
+		p: p,
 		healthCheck: prometheus.NewGauge(prometheus.GaugeOpts{
 			Namespace: namespace,
 			Subsystem: stateSubsystem,
@@ -55,16 +64,42 @@ func newStateMetrics() *stateMetrics {
 	}
 }
 
-func (m stateMetrics) register() {
-	prometheus.MustRegister(m.healthCheck)
-}
-
-func (m stateMetrics) unregister() {
+func (m *stateMetrics) unregister() {
 	prometheus.Unregister(m.healthCheck)
 }
 
-func (m stateMetrics) SetHealth(s int32) {
+func (m *stateMetrics) SetHealth(s int32) {
 	m.healthCheck.Set(float64(s))
+}
+
+func (m *stateMetrics) updateHealthStatus() {
+	ctx, cancel := context.WithTimeout(context.Background(), checkNeoFSConnectionTimeOut)
+	defer cancel()
+
+	cl, err := m.p.RawClient()
+	// Only "no healthy client" error is possible.
+	if err != nil {
+		m.healthCheck.Set(0)
+		return
+	}
+
+	// Check actual connection is really alive.
+	_, err = cl.NetworkInfo(ctx, client.PrmNetworkInfo{})
+	if err != nil && strings.Contains(err.Error(), "connection error") {
+		m.healthCheck.Set(0)
+		return
+	}
+
+	m.healthCheck.Set(1)
+}
+
+func (m *stateMetrics) Collect(ch chan<- prometheus.Metric) {
+	m.updateHealthStatus()
+	m.healthCheck.Collect(ch)
+}
+
+func (m stateMetrics) Describe(descs chan<- *prometheus.Desc) {
+	m.healthCheck.Describe(descs)
 }
 
 // NewPrometheusService creates a new service for gathering prometheus metrics.