+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This document outlines major changes between releases.
 
 ### Changed
 - Default polling interval for container operations from fixed 1s to dynamic value based on block time (#1194)
+- `neofs_s3_gw_state_health` metric changes the state according to the gate health status (#1203)
 
 ### Fixed
 - Incorrect EpochDuration processing wrt API 2.18 leading to improper expiration calculations (#1193)